Moving the existing file upload code into subdirectory so it's easier to find things and doesn't get mangled in with everything else.  Had to update some paths to get it to work again in a few places.
