+++ app/javascript/components/FileUpload/FileList/BadList/BadList.js
