+++ app/javascript/components/FileUpload/FileList/File/File.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import {TabularCheckStatus} from "../../../containers/UploadFiles";
+import {TabularCheckStatus} from "../../../../containers/FileUpload/UploadFiles";
 
-import ellipsize from '../../../lib/string_patch';
+import ellipsize from '../../../../lib/string_patch';
 
 import classes from './File.module.css';
