+++ app/javascript/components/FileUpload/Instructions/Instructions.js
