+++ app/javascript/components/FileUpload/UploadType/UploadType.js
