+++ app/javascript/components/FileUpload/ValidateFiles/ValidateFiles.js
