+++ app/javascript/components/FileUpload/WarningMessage/WarningMessage.js
