+++ stash/stash_engine/app/views/stash_engine/resources/upload.html.erb
@@ -12,7 +12,7 @@
 <% end %>
 
 <!-- map(&:attributes) one way for translating ActiveRecord field type to json at React side adequately -->
-<%= react_component('UploadFiles', {
+<%= react_component('FileUpload/UploadFiles', {
   resource_id: @resource.id,
   file_uploads: @resource.generic_files.validated_table.as_json(
     methods: :type, include: { frictionless_report: { only: [:report, :status] } }
